fix: must be restarted to take effect after modify server url
